Return values for density, DOJ, and norm. Hamming distance
The special data types `Fraction` and `Decimal` did not yield a relevant precision gain in any practical purpose; however, at least `Fraction` complicated data handling when evaluating experiments.

+++ taupy/basic/core.py
@@ -1,4 +1,3 @@
-from decimal import Decimal
 from itertools import combinations
 from math import log2
 from sympy.logic import (And, Implies, Not)
@@ -89,7 +88,7 @@ class Base():
         ([Betz2013]_ , pp. 44–49).
         """
         sigma = satisfiability_count(self)
-        return Decimal((len(self.atoms()) - log2(sigma)) / len(self.atoms()))
+        return (len(self.atoms()) - log2(sigma)) / len(self.atoms())
     
     def list_of_premises(self):
         """
